Fix failures due to invalid numpy scalars. (#634)
* Fix failures due to invalid numpy scalars.

* Remove numpy pins.

* Update dev dependency installation invocation.

* Manual numpy upgrade.

+++ deepcell/image_generators/tracking.py
@@ -218,8 +218,8 @@ class SiameseIterator(Iterator):
         self.min_track_length = min_track_length
         self.features = sorted(features)
         self.sync_transform = sync_transform
-        self.neighborhood_scale_size = np.int(neighborhood_scale_size)
-        self.neighborhood_true_size = np.int(neighborhood_true_size)
+        self.neighborhood_scale_size = np.int_(neighborhood_scale_size)
+        self.neighborhood_true_size = np.int_(neighborhood_true_size)
         self.image_data_generator = image_data_generator
         self.data_format = data_format
         self.save_to_dir = save_to_dir
@@ -372,7 +372,7 @@ class SiameseIterator(Iterator):
         y_padded = np.pad(y_frame, pads, mode='constant', constant_values=0)
         props = regionprops(np.squeeze(np.int32(y_padded == cell_label)))
         center_x, center_y = props[0].centroid
-        center_x, center_y = np.int(center_x), np.int(center_y)
+        center_x, center_y = np.int_(center_x), np.int_(center_y)
         X_reduced = X_padded[
             center_x - self.neighborhood_true_size:center_x + self.neighborhood_true_size,
             center_y - self.neighborhood_true_size:center_y + self.neighborhood_true_size,
@@ -904,7 +904,7 @@ class SiameseIterator(Iterator):
 
             if type_cell == 2:
                 # There should always be 2 daughters but not always a valid label
-                label_2 = np.int(np.random.choice(track_id['daughters']))
+                label_2 = np.int_(np.random.choice(track_id['daughters']))
                 daughter_track = self.reverse_track_ids[batch][label_2]
                 frame_2 = np.amin(self.track_ids[daughter_track]['frames'])
                 frames_2 = [frame_2]
